zephyr: cpu: add cpu.c in zephyr_library_sources
This file contains an implementation of cpu management code using the
functionality provided by the zephyr kernel.

Initial code contains implementation of cpu_disable_core(). The function
forces the PM_STATE_SOFT_OFF state with the zephyr power manager. It
then waits for the core to report its deactivation and turn it off.

Signed-off-by: Tomasz Leman <tomasz.m.leman@intel.com>